+++ CHANGES.rst
@@ -5,7 +5,8 @@ Changelog
 1.0a2 (unreleased)
 ------------------
 
-- Nothing changed yet.
+- Fix an issue with nested optgroup.
+  [mpeeters]
 
 
 1.0a1 (2016-07-19)